Split off scrollable panel
